robot_ki_dashboard.py
import streamlit as st
import os
import sys
import platform
import psutil
import datetime

st.set_page_config(
    page_title="🤖 ROBOTER KI APP - Dashboard Übersicht", page_icon="🤖", layout="wide"
)

st.title("🤖 ROBOTER KI APP - Dashboard Übersicht")
st.markdown("---")

# System Information
st.header("🖥️ System Information")
col1, col2, col3 = st.columns(3)

with col1:
    st.metric("Python Version", sys.version.split()[0])
    st.metric("Platform", platform.system())

with col2:
    st.metric("CPU Cores", psutil.cpu_count())
    st.metric("RAM Total", f"{psutil.virtual_memory().total / (1024**3):.1f} GB")

with col3:
    st.metric("Current Time", datetime.datetime.now().strftime("%H:%M:%S"))
    st.metric("Date", datetime.datetime.now().strftime("%d.%m.%Y"))

st.markdown("---")

# Cache Status
st.header("🗂️ Cache Status")
st.success("✅ `__pycache__` Verzeichnis wurde erfolgreich gelöscht")
st.info("Alle kompilierten Python-Bytecode-Dateien (.pyc) wurden entfernt.")

st.markdown(
    """
**Vorteile der Cache-Bereinigung:**
- 🔄 **Neukompilierung**: Alle Module werden beim nächsten Start frisch kompiliert
- 🐛 **Stale Bytecode**: Vermeidung von Problemen mit veraltetem Bytecode
- 📦 **Import-Fehler**: Behebung von Modul-Import-Problemen
- 🔧 **Entwicklung**: Sicherstellung, dass Änderungen an .py-Dateien wirksam werden
"""
)

st.markdown("---")

# Module Overview
st.header("📚 Module Übersicht")
modules_path = r"c:\Users\nazmi\modules"

if os.path.exists(modules_path):
    py_files = [f for f in os.listdir(modules_path) if f.endswith(".py")]
    st.metric("Python Module", len(py_files))

    # Show some example modules
    if py_files:
        st.subheader("Beispiel-Module:")
        cols = st.columns(4)
        for i, module in enumerate(py_files[:8]):  # Show first 8
            with cols[i % 4]:
                st.code(module, language="python")
else:
    st.error("Modules-Verzeichnis nicht gefunden")

# Integration Status
st.header("🔗 Projekt-Integrationen")
st.success("✅ Desktop-Tutorial (AethelosGAZI) integriert")
st.info("Integrierte Projekte:")
integrations = [
    "ZenithCoreSystem - Autonomous Zenith Optimizer",
    "Kontrollturm - System Control Center",
    "MegaUltraNetwork - AI Network Hub",
    "zenithapi - Zenith REST API",
    "AI_CORE - Core AI Integrator",
    "DesktopTutorial - AethelosGAZI System",
]

for integration in integrations:
    st.write(f"• {integration}")

st.markdown("---")

# Actions
st.header("🎯 Aktionen")
col1, col2, col3 = st.columns(3)

with col1:
    if st.button("🔄 System Neustart", type="primary"):
        st.success("System würde hier neu gestartet werden...")

with col2:
    if st.button("📊 Performance Check"):
        st.info("Performance-Check würde hier ausgeführt werden...")

with col3:
    if st.button("🧹 Cache leeren"):
        st.success("Cache wurde bereits geleert!")

st.markdown("---")

# Footer
st.markdown(
    """
**🤖 ROBOTER KI APP Dashboard**
- Automatische Systemüberwachung
- Modul-Management
- Performance-Optimierung
- Quantum-Integration
"""
)

st.caption(
    "Dashboard erstellt am: " + datetime.datetime.now().strftime("%d.%m.%Y %H:%M:%S")
)
