requirements.txt
torch
torchvision
torchaudio
openai
python-dotenv
streamlit
pandas
modin
dask
boto3
cryptography
gtts
SpeechRecognition

# Additional dependencies for Maximum Autonomous Profit System
fastapi>=0.104.0
uvicorn>=0.24.0
aiohttp>=3.9.0
numpy>=1.24.0
plotly>=5.17.0
requests>=2.31.0
httpx>=0.25.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
scipy>=1.11.0
scikit-learn>=1.3.0
sqlalchemy>=2.0.0
redis>=5.0.0
ccxt>=4.1.0
forex-python>=1.8
pyowm>=3.3.0
textblob>=0.17.1
vaderSentiment>=3.3.2
beautifulsoup4>=4.12.0
Pillow>=10.0.0